Refactor Stripe adapter: move helper functions to dto package
