+++ payments/internal/dto/money.go
@@ -1,4 +1,4 @@
-package stripeadp
+package dto
 
 import (
 	"github.com/stripe/stripe-go/v82"
@@ -6,7 +6,7 @@ import (
 	"google.golang.org/genproto/googleapis/type/money"
 )
 
-// fromMinor converts minor currency units to money.Money.
-func fromMinor(cur stripe.Currency, v int64) *money.Money {
+// FromMinor converts minor currency units to money.Money.
+func FromMinor(cur stripe.Currency, v int64) *money.Money {
 	return ledger.MinorUnitsToAmount(string(cur), v)
 }
